+++ sql/core/src/main/java/org/apache/spark/sql/execution/datasources/parquet/VectorizedColumnReader.java
@@ -32,6 +32,7 @@ import org.apache.parquet.schema.OriginalType;
 import org.apache.parquet.schema.PrimitiveType;
 
 import org.apache.spark.sql.catalyst.util.DateTimeUtils;
+import org.apache.spark.sql.execution.datasources.SchemaColumnConvertNotSupportedException;
 import org.apache.spark.sql.execution.vectorized.WritableColumnVector;
 import org.apache.spark.sql.types.DataTypes;
 import org.apache.spark.sql.types.DecimalType;
@@ -235,10 +236,10 @@ public class VectorizedColumnReader {
   /**
    * Helper function to construct exception for parquet schema mismatch.
    */
-  private ParquetSchemaColumnConvertNotSupportedException constructConvertNotSupportedException(
+  private SchemaColumnConvertNotSupportedException constructConvertNotSupportedException(
       ColumnDescriptor descriptor,
       WritableColumnVector column) {
-    return new ParquetSchemaColumnConvertNotSupportedException(
+    return new SchemaColumnConvertNotSupportedException(
       Arrays.toString(descriptor.getPath()),
       descriptor.getType().toString(),
       column.dataType().toString());
